+++ domain/client/src/main/kotlin/io/github/siyual_park/client/property/RootClientProperty.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.client.configuration
+package io.github.siyual_park.client.property
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
