+++ module/data/src/main/kotlin/io/github/siyual_park/data/property/MigrationProperty.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.data.configuration
+package io.github.siyual_park.data.property
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
